주문생성까지 완성한 거 올림 item쪽 단가 또는 가격에 타입이 다르다고 해서 그거 수정 잘 확인

+++ src/main/java/com/t1membership/item/domain/ItemEntity.java
@@ -5,6 +5,8 @@ import com.t1membership.item.constant.*;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Builder
@@ -23,7 +25,7 @@ public class ItemEntity extends BaseEntity {
     private String itemName;
 
     @Column(name = "item_price", nullable = false)
-    private int itemPrice;
+    private BigDecimal itemPrice;
 
     @Column(name = "item_stock", nullable = false)
     private int itemStock;
